added e_obj and d_dot_e_normal to Ray
removed thread stuff from qJuliaSet.
TODO: change objects to std::vector<SceneObject>, improve qJuliaSet
doesIntersect t check to include epsilons

+++ Src/thread-pool.cc
@@ -101,7 +101,7 @@ void ThreadPool::dispatcher() {
   }
 }
 
-void ThreadPool::worker(int workerID) {
+void ThreadPool::worker(size_t workerID) {
   while (true) {
     // wait to gain exclusive access to dispatcher's "office", wait for a task
     // to be provided, then leave the dispatcher's "office"
